requirements.txt
# CECAN Platform - Python Dependencies
# Backend Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.0.0

# Database ORM
sqlalchemy>=2.0.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# PDF Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# AI & ML
google-generativeai>=0.3.0
numpy>=1.24.0
thefuzz>=0.19.0

# Environment & Config
python-dotenv>=1.0.0

# Optional but recommended
# For production deployment
gunicorn>=21.0.0  # Production WSGI server

# For testing
pytest>=7.4.0
httpx>=0.25.0  # For FastAPI testing

# For monitoring
python-multipart>=0.0.6
# Data Analysis & Export
pandas>=2.0.0
openpyxl>=3.1.0
